adds FeaturedProductsHeader and ProductCard component
